API Change:
RBQRealmNotificationManager now uses a different scheme to pass changes to any listeners. Instead of passing arrays of added/deleted/inserted RBQSafeRealmObjects that are not sorted by entity type, the notification manager now passes a NSDictionary that is keyed by the class name of the entity type. The object in the dictionary is an instance of RBQEntityChangesObject which contain the added/deleted/inserted arrays for only that entity.

This improves performance when tracking changes from many entity types for different listeners. A listener simply needs to check if there is an object for the class name key it is looking for, and operate on those changes versus enumerating all of the chanegs to find the appropriate ones.

Other:
SLowly improving the internal documentation to be in line with appledoc.

+++ RBQFetchedResultsControllerExample/RBQFetchedResultsControllerExample/ExampleTableViewController.m
@@ -213,7 +213,7 @@ forRowAtIndexPath:(NSIndexPath *)indexPath
 - (void)controllerDidChangeContent:(RBQFetchedResultsController *)controller
 {
     NSLog(@"Ending updates");
-    NSLog(@"Fetched %ld Items After Change", self.fetchedResultsController.fetchedObjects.count);
+    NSLog(@"Fetched %ld Items After Change", (unsigned long)self.fetchedResultsController.fetchedObjects.count);
     @try {
         [self.tableView endUpdates];
     }
@@ -233,7 +233,7 @@ forRowAtIndexPath:(NSIndexPath *)indexPath
     
     NSLog(@"DID BEGIN DELETE");
     
-    NSLog(@"Fetched %ld Items Before Delete", self.fetchedResultsController.fetchedObjects.count);
+    NSLog(@"Fetched %ld Items Before Delete", (unsigned long)self.fetchedResultsController.fetchedObjects.count);
     
     // Test deleting a section (comment out above to test)
     RLMResults *objectInFirstSection = [TestObject objectsWhere:@"%K == %@",@"sectionName",@"First Section"];
@@ -250,7 +250,7 @@ forRowAtIndexPath:(NSIndexPath *)indexPath
 - (IBAction)didClickInsertButton:(UIBarButtonItem *)sender
 {
     NSLog(@"DID BEGIN INSERT");
-    NSLog(@"Fetched %ld Items Before Insert", self.fetchedResultsController.fetchedObjects.count);
+    NSLog(@"Fetched %ld Items Before Insert", (unsigned long)self.fetchedResultsController.fetchedObjects.count);
     [self insertObject];
     NSLog(@"DID END INSERT");
 }
